Organize the Search Handler classes and models and write tests
